refactor(logging): move CorrelationFilter to http package
