+++ api/src/main/java/com/reliaquest/api/http/CorrelationFilter.java
@@ -1,4 +1,4 @@
-package com.reliaquest.api.log;
+package com.reliaquest.api.http;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
