+++ api/src/test/java/com/reliaquest/api/http/CorrelationFilterTest.java
@@ -1,8 +1,9 @@
-package com.reliaquest.api.log;
+package com.reliaquest.api.http;
 
 import com.reliaquest.api.controller.EmployeeController;
 import com.reliaquest.api.controller.EmployeeControllerAdvice;
 import com.reliaquest.api.exception.DownstreamUnavailableException;
+import com.reliaquest.api.http.CorrelationFilter;
 import com.reliaquest.api.model.Employee;
 import com.reliaquest.api.service.EmployeeService;
 import org.junit.jupiter.api.Assertions;
